requirements.txt
beautifulsoup4
requests
google-genai
browser-use
langchain_google_genai